move models to module
