--- ngse/models.py
@@ -1,209 +1,0 @@
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.schema import Table
-from sqlalchemy.orm import relationship
-from sqlalchemy.schema import Table
-from sqlalchemy.dialects.postgresql import ARRAY, JSON
-from sqlalchemy import (
-	Column,
-	Integer,
-	Text,
-	Boolean,
-	DateTime,
-	ForeignKey,
-	func
-)
-import bcrypt
-
-Base = declarative_base()
-
-form_category_association = Table('form_category_association', Base.metadata,
-	Column('form_types_id', Integer, ForeignKey('form_types.id')),
-	Column('categories_id', Integer, ForeignKey('categories.id'))
-)
-
-class FormType(Base):
-	__tablename__ = 'form_types'
-
-	id = Column(Integer, primary_key=True)
-	name = Column(Text, nullable=False)
-	page_sequence = Column(ARRAY(Integer))
-	date_created = Column(DateTime, nullable=False, server_default=func.now())
-	last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	forms = relationship("Form", back_populates="form_type") # child relationship
-	categories = relationship("Category", secondary=form_category_association, back_populates="form_type")
-	user_type_id = Column(Integer, ForeignKey('user_types.id'))
-
-	def as_dict(self):
-		return {c.name: str(getattr(self, c.name)) for c in self.__table__.columns}
-
-class Form(Base):
-	__tablename__ = 'forms'
-
-	id = Column(Integer, primary_key=True)
-	name = Column(Text, nullable=False)
-	date_created = Column(DateTime, nullable=False, server_default=func.now())
-	last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	date_start = Column(DateTime, nullable=False)
-	date_end = Column(DateTime, nullable=False)
-
-	form_type_id = Column(Integer, ForeignKey('form_types.id')) # parent
-	form_type = relationship("FormType", back_populates="forms") # parent relationship
-
-	def as_dict(self):
-		return {c.name: str(getattr(self, c.name)) for c in self.__table__.columns}
-
-class Category(Base):
-	__tablename__ = 'categories'
-
-	id = Column(Integer, primary_key=True)
-	name = Column(Text, nullable=False)
-	date_created = Column(DateTime, nullable=False, server_default=func.now())
-	last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	form_type = relationship("FormType", secondary=form_category_association, back_populates="categories") # parent relationship
-
-	elements = relationship("Element", back_populates="category")
-
-	def as_dict(self):
-		return {c.name: str(getattr(self, c.name)) for c in self.__table__.columns}
-
-class Element(Base):
-	__tablename__ = 'elements'
-
-	id = Column(Integer, primary_key=True)
-	name = Column(Text, nullable=False)
-	date_created = Column(DateTime, nullable=False, server_default=func.now())
-	last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	category_id = Column(Integer, ForeignKey('categories.id')) # parent
-	category = relationship("Category", back_populates="elements") #parent relationship
-
-	# form_type_id = Column(Integer, ForeignKey('form_types.id')) # parent
-	# form_type = relationship("FormType", back_populates="questions") # parent relationship
-
-	text = Column(Text, nullable=False)
-	klass = Column(Text, nullable=False, default='question')
-	kind = Column(Text, nullable=False, default='text')
-	width = Column(Integer, nullable=False, default=16)
-	choices = Column(ARRAY(Text))
-	default = Column(Text)
-	required = Column(Boolean)
-
-	meta = Column(JSON)
-
-	answers = relationship("Answer", back_populates="element")
-
-	def as_dict(self):
-		return {c.name: str(getattr(self, c.name)) for c in self.__table__.columns}
-
-class Answer(Base):
-	__tablename__ = 'answers'
-
-	id = Column(Integer, primary_key=True)
-	text = Column(Text, nullable=False)
-	date_created = Column(DateTime, nullable=False, server_default=func.now())
-	last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	element_id = Column(Integer, ForeignKey('elements.id')) # parent
-	element = relationship("Element", back_populates='answers') # parent relationship
-
-	user_id = Column(Integer, ForeignKey('users.id')) # parent
-	user = relationship("User", back_populates='answers') # parent relationship
-
-	def as_dict(self):
-		return {c.name: str(getattr(self, c.name)) for c in self.__table__.columns}
-
-class UserType(Base):
-	__tablename__ = 'user_types'
-
-	id = Column(Integer, primary_key=True)
-	name = Column(Text, nullable=False)
-	# date_created = Column(DateTime, nullable=False, server_default=func.now())
-	# last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	user = relationship("User", back_populates='user_type')
-
-	def as_dict(self):
-		return {c.name: str(getattr(self, c.name)) for c in self.__table__.columns}
-
-class User(Base):
-	__tablename__ = 'users'
-
-	id = Column(Integer, primary_key=True)
-	name = Column(Text, nullable=False)
-	date_created = Column(DateTime, nullable=False, server_default=func.now())
-	last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	email = Column(Text, nullable=False)
-	password = Column(Text, nullable=False)
-	
-	submitted = Column(Boolean, default=False)
-	# meta = Column(JSON) #status can be submitted, on process, accepted, rejected
-
-	user_type_id = Column(Integer, ForeignKey('user_types.id'), default=4)
-	user_type = relationship("UserType", back_populates='user')
-
-	# applicant_attr = relationship("ApplicantAttribute", back_populates='applicant')
-	# rec_a = relationship("ApplicantAttribute", back_populates='recommenderA')
-	# rec_b = relationship("ApplicantAttribute", back_populates='recommenderB')
-	# rec_c = relationship("ApplicantAttribute", back_populates='recommenderC')
-
-	# applicant_attr = relationship("ApplicantAttribute")
-	answers = relationship("Answer", back_populates="user")
-
-class CategoryStatus(Base):
-	__tablename__ = 'category_status'
-
-	id = Column(Integer, primary_key=True)
-	date_created = Column(DateTime, nullable=False, server_default=func.now())
-	last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	status = Column(Boolean, nullable=False, default=False)
-
-	user_id = Column(Integer, ForeignKey('users.id')) # parent
-	# user = relationship("User", back_populates='answers') # parent relationship
-
-	category_id = Column(Integer, ForeignKey('categories.id')) # parent
-	# category = relationship("Category", back_populates="elements") #parent relationship
-
-class ApplicantAttribute(Base):
-	__tablename__ = 'applicant_attrs'
-
-	id = Column(Integer, primary_key=True)
-	date_created = Column(DateTime, nullable=False, server_default=func.now())
-	last_modified = Column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
-
-	scholarship = Column(Boolean, nullable=False, default=False)
-	application_status = Column(Text, nullable=False, default='None')
-	validation_status = Column(Text, nullable=False, default='None')
-
-	level = Column(Text)
-	program = Column(Text)
-	program_type = Column(Text)
-	student_type = Column(Text)
-	choice_1 = Column(Text)
-	choice_2 = Column(Text)
-	choice_3 = Column(Text)
-	adviser = Column(Text)
-
-	start_of_study = Column(Text)
-	year = Column(Text)
-	other_scholarship = Column(Text)
-	other_scholarship_name = Column(Text)
-
-	answered_pos = Column(Boolean, default=False)
-
-	recommender_a = Column(Integer, ForeignKey('users.id'))
-	recommender_b = Column(Integer, ForeignKey('users.id'))
-	
-	recommender_c = Column(Integer, ForeignKey('users.id'))
-
-	applicant_id = Column(Integer, ForeignKey('users.id'))
-	# applicant = relationship("User", back_populates='applicant_attr')
-
-	applicant = relationship("User", foreign_keys='ApplicantAttribute.applicant_id')
-	rec_a = relationship("User", foreign_keys='ApplicantAttribute.recommender_a')
-	rec_b = relationship("User", foreign_keys='ApplicantAttribute.recommender_b')
-	rec_c = relationship("User", foreign_keys='ApplicantAttribute.recommender_c')
